simplify reply/notification API
