+++ examples/with-async-std/src/main.rs
@@ -2,14 +2,14 @@
 #![deny(clippy::unimplemented, clippy::todo)]
 
 use polyfuse::{
-    bytes::{write_bytes, Bytes},
     op,
-    reply::{AttrOut, EntryOut, FileAttr, ReaddirOut, Reply},
+    reply::{AttrOut, EntryOut, FileAttr, ReaddirOut},
     Config, MountOptions, Operation, Request, Session,
 };
 
-use anyhow::Context as _;
-use std::{io, os::unix::prelude::*, path::PathBuf, time::Duration};
+use anyhow::{ensure, Context as _, Result};
+use async_std::task::{self, JoinHandle};
+use std::{io, os::unix::prelude::*, path::PathBuf, sync::Arc, time::Duration};
 
 const TTL: Duration = Duration::from_secs(60 * 60 * 24 * 365);
 const ROOT_INO: u64 = 1;
@@ -18,29 +18,33 @@ const HELLO_FILENAME: &str = "hello.txt";
 const HELLO_CONTENT: &[u8] = b"Hello, world!\n";
 
 #[async_std::main]
-async fn main() -> anyhow::Result<()> {
+async fn main() -> Result<()> {
     tracing_subscriber::fmt::init();
 
     let mut args = pico_args::Arguments::from_env();
 
     let mountpoint: PathBuf = args.free_from_str()?.context("missing mountpoint")?;
-    anyhow::ensure!(mountpoint.is_dir(), "mountpoint must be a directory");
+    ensure!(mountpoint.is_dir(), "mountpoint must be a directory");
 
     let session =
         AsyncSession::mount(mountpoint, MountOptions::default(), Config::default()).await?;
 
-    let fs = Hello::new();
+    let fs = Arc::new(Hello::new());
 
     while let Some(req) = session.next_request().await? {
-        let reply = ReplyWriter { req: &req };
+        let fs = fs.clone();
 
-        match req.operation()? {
-            Operation::Lookup(op) => fs.lookup(op, reply).await?,
-            Operation::Getattr(op) => fs.getattr(op, reply).await?,
-            Operation::Read(op) => fs.read(op, reply).await?,
-            Operation::Readdir(op) => fs.readdir(op, reply).await?,
-            _ => reply.error(libc::ENOSYS)?,
-        }
+        let _: JoinHandle<Result<()>> = task::spawn(async move {
+            match req.operation()? {
+                Operation::Lookup(op) => fs.lookup(&req, op).await?,
+                Operation::Getattr(op) => fs.getattr(&req, op).await?,
+                Operation::Read(op) => fs.read(&req, op).await?,
+                Operation::Readdir(op) => fs.readdir(&req, op).await?,
+                _ => req.reply_error(libc::ENOSYS)?,
+            }
+
+            Ok(())
+        });
     }
 
     Ok(())
@@ -101,7 +105,7 @@ impl Hello {
         attr.gid(self.gid);
     }
 
-    async fn lookup(&self, op: op::Lookup<'_>, reply: ReplyWriter<'_>) -> io::Result<()> {
+    async fn lookup(&self, req: &Request, op: op::Lookup<'_>) -> io::Result<()> {
         match op.parent() {
             ROOT_INO if op.name().as_bytes() == HELLO_FILENAME.as_bytes() => {
                 let mut out = EntryOut::default();
@@ -109,31 +113,30 @@ impl Hello {
                 out.ino(HELLO_INO);
                 out.ttl_attr(TTL);
                 out.ttl_entry(TTL);
-                reply.ok(out)
+                req.reply(out)
             }
-            _ => reply.error(libc::ENOENT),
+            _ => req.reply_error(libc::ENOENT),
         }
     }
 
-    async fn getattr(&self, op: op::Getattr<'_>, reply: ReplyWriter<'_>) -> io::Result<()> {
-        let fill_attr = match op.ino() {
-            ROOT_INO => Self::fill_root_attr,
-            HELLO_INO => Self::fill_hello_attr,
-            _ => return reply.error(libc::ENOENT),
+    async fn getattr(&self, req: &Request, op: op::Getattr<'_>) -> io::Result<()> {
+        let mut out = AttrOut::default();
+        match op.ino() {
+            ROOT_INO => self.fill_root_attr(out.attr()),
+            HELLO_INO => self.fill_hello_attr(out.attr()),
+            _ => return req.reply_error(libc::ENOENT),
         };
 
-        let mut out = AttrOut::default();
-        fill_attr(self, out.attr());
         out.ttl(TTL);
 
-        reply.ok(out)
+        req.reply(out)
     }
 
-    async fn read(&self, op: op::Read<'_>, reply: ReplyWriter<'_>) -> io::Result<()> {
+    async fn read(&self, req: &Request, op: op::Read<'_>) -> io::Result<()> {
         match op.ino() {
             HELLO_INO => (),
-            ROOT_INO => return reply.error(libc::EISDIR),
-            _ => return reply.error(libc::ENOENT),
+            ROOT_INO => return req.reply_error(libc::EISDIR),
+            _ => return req.reply_error(libc::ENOENT),
         }
 
         let mut data: &[u8] = &[];
@@ -145,7 +148,7 @@ impl Hello {
             data = &data[..std::cmp::min(data.len(), size)];
         }
 
-        reply.ok(data)
+        req.reply(data)
     }
 
     fn dir_entries(&self) -> impl Iterator<Item = (u64, &DirEntry)> + '_ {
@@ -155,9 +158,9 @@ impl Hello {
         })
     }
 
-    async fn readdir(&self, op: op::Readdir<'_>, reply: ReplyWriter<'_>) -> io::Result<()> {
+    async fn readdir(&self, req: &Request, op: op::Readdir<'_>) -> io::Result<()> {
         if op.ino() != ROOT_INO {
-            return reply.error(libc::ENOTDIR);
+            return req.reply_error(libc::ENOTDIR);
         }
 
         let mut out = ReaddirOut::new(op.size() as usize);
@@ -174,24 +177,7 @@ impl Hello {
             }
         }
 
-        reply.ok(out)
-    }
-}
-
-struct ReplyWriter<'req> {
-    req: &'req Request,
-}
-
-impl ReplyWriter<'_> {
-    fn ok<T>(self, arg: T) -> io::Result<()>
-    where
-        T: Bytes,
-    {
-        write_bytes(self.req, Reply::new(self.req.unique(), 0, arg))
-    }
-
-    fn error(self, code: i32) -> io::Result<()> {
-        write_bytes(self.req, Reply::new(self.req.unique(), code, ()))
+        req.reply(out)
     }
 }
